refactor: 위도, 경도 자료형 수정

+++ src/main/java/com/team01/billage/product/dto/ProductDetailResponseDto.java
@@ -2,7 +2,6 @@ package com.team01.billage.product.dto;
 
 import com.team01.billage.category.dto.CategoryProductResponseDto;
 import com.team01.billage.product_review.dto.ShowReviewResponseDto;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -25,8 +24,8 @@ public class ProductDetailResponseDto { // 상품 상세 확인 DTO
     private String rentalStatus;
     private int dayPrice;
     private Integer weekPrice;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
     private int viewCount;
     private LocalDateTime updatedAt;
     private ProductSellerResponseDto seller;
